refactor: refactored the components
